+++ pkgs/ffigen/lib/src/config_provider/config_types.dart
@@ -14,6 +14,7 @@ import 'package:quiver/pattern.dart' as quiver;
 import '../code_generator.dart';
 import 'config.dart';
 import 'path_finder.dart';
+import 'utils.dart';
 
 export 'package:pub_semver/pub_semver.dart' show Version;
 
@@ -382,10 +383,29 @@ class YamlMemberIncluder {
 List<String> defaultCompilerOpts(
   Logger logger, {
   bool macIncludeStdLib = true,
+  bool cpp = false,
+}) => cpp
+    ? [
+        '-x',
+        'c++',
+        '-std=c++17',
+        if (Platform.isMacOS) ...['-isysroot', macSdkPath],
+      ]
+    : [
+        if (Platform.isMacOS && macIncludeStdLib)
+          ...getCStandardLibraryHeadersForMac(logger),
+        if (Platform.isMacOS) '-Wno-nullability-completeness',
+      ];
+
+/// Computes compiler options based on [config] and [logger].
+List<String> computeCompilerOpts({
+  required FfiGenerator config,
+  required Logger logger,
 }) => [
-  if (Platform.isMacOS && macIncludeStdLib)
-    ...getCStandardLibraryHeadersForMac(logger),
-  if (Platform.isMacOS) '-Wno-nullability-completeness',
+  if (config.input.appendCompilerOptions ||
+      config.input.compilerOptions == null)
+    ...defaultCompilerOpts(logger, cpp: config.cpp != null),
+  if (config.input.compilerOptions != null) ...config.input.compilerOptions!,
 ];
 
 /// Handles config for automatically added compiler options.
@@ -437,8 +457,20 @@ class FfiNativeConfig {
   const FfiNativeConfig({required this.enabled, this.assetId});
 }
 
+/// Configuration for generating a symbol file.
+///
+/// Symbol files allow other FFIgen runs to import and reuse symbols defined in
+/// these bindings rather than regenerating them.
 class SymbolFile {
+  /// The package or file URI that other bindings will use to import the
+  /// generated Dart bindings for these symbols.
+  ///
+  /// Using a `package:...` URI is recommended for cross-package imports, so
+  /// that other packages can resolve the import regardless of directory
+  /// structure.
   final Uri importPath;
+
+  /// The file URI where YAML symbol file will be generated.
   final Uri output;
 
   SymbolFile(this.importPath, this.output);