+++ pkgs/ffigen/lib/src/context.dart
@@ -27,8 +27,7 @@ class Context {
   bool hasSourceErrors = false;
   final reportedCommentRanges = <((String, int), (String, int))>{};
   final libs = LibraryImports();
-  late final compilerOpts =
-      config.input.compilerOptions ?? defaultCompilerOpts(logger);
+  late final compilerOpts = computeCompilerOpts(config: config, logger: logger);
   final Scope rootScope = Scope.createRoot('root');
   final Scope rootObjCScope = Scope.createRoot('objc_root');
   late final ExtraSymbols extraSymbols;