+++ pkgs/ffigen/CHANGELOG.md
@@ -1,6 +1,9 @@
-## 22.0.1-wip
+## 22.1.0-wip
 
 - The YAML configuration format is deprecated and will be removed in a future release.
+- Added `Input.appendCompilerOptions` to allow appending custom compiler
+  options to default options instead of replacing them.
+- Default to C++17 compiler options when C++ bindings are enabled.
 - Preserve documentation comments on same-name typedefs when generating the
   underlying type binding.
 - Fix [a bug](https://github.com/dart-lang/native/issues/3592) where functions
@@ -14,6 +17,7 @@
   canonical type.
 - Generate bindings for declarations inside C++ `extern "C" { ... }`
   blocks, which were previously skipped entirely.
+- Allow `package:cli_util` versions `0.5.x` and `0.6.x`.
 
 ## 22.0.0